Add initial changes for api work
This includes:
-Create files for api curl scripts
-Move around some files to keep everything organized
-Start functions in events.js
