+++ assets/scripts/games/update.sh
@@ -1,0 +1,7 @@
+curl --include --request PATCH "https://ga-library-api.herokuapp.com/books/${ID}" \
+  --header "Content-type: application/json" \
+  --data '{
+    "book": {
+      "title": "'"${TITLE}"'"
+    }
+  }'
