+++ index.html
@@ -85,11 +85,15 @@
             <h2 id="message"></h2>
           </div>
         </div>
-        <div class="col-xs-4 new-game">
+        <div class="col-xs-3 new-game" id="new-game">
           <br>
           <button type="submit" class="new-game-button">New Game</button>
         </div>
-        <div class="col-xs-4 change-pswd">
+        <div class="col-xs-3 stats">
+          <br>
+          <button type="submit" class="get-games-button" id="get-games">Statistics</button>
+        </div>
+        <div class="col-xs-3 change-pswd">
           <br>
           <button type="submit" class="trigger-modal-button" id="trigger-modal-button">Account Info</button>
           <div id="modal" class="modal">
@@ -107,7 +111,7 @@
             </div>
           </div>
         </div>
-        <div class="col-xs-4 sign-out">
+        <div class="col-xs-3 sign-out">
           <form id="sign-out">
             <br>
             <input type="submit" name="submit" class="sign-out-button" value="Sign Out">
